Add ability to extract event information
This changes naming conventions to call the date "date" and use the
term "event" to refer to metadata about the flight.

Any file ending with the date or with any case of _A, _primary, is
considered the primary flight and stored the event is returned as
"primary". For all other cases the event is whatever comes after the '_'
after the date.
